feat(PHP): Don't delete the group when is defined to is_locked to true
When I remove a space, I made sure that the group to delete has its is_locked value defined to false. It means this group is LDAP or other than database.

+++ lib/Controller/WorkspaceController.php
@@ -345,8 +345,10 @@ class WorkspaceController extends Controller {
         $groups = [];
     	$this->logger->debug('Removing workspaces groups.');
         foreach ( array_keys($workspace['groups']) as $group ) {
-            $groups[] = $group;
-            $this->groupManager->get($group)->delete();
+            if (!$workspace['groups'][$group]['is_locked']) {
+                $groups[] = $group;
+                $this->groupManager->get($group)->delete();
+            }
         }
 
 	    return new JSONResponse([
